pyproject.toml
[project]
name = "invoice-intake-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "asyncio>=4.0.0",
    "openai-agents>=0.9.1",
    "pdf2image>=1.17.0",
    "pdfminer>=20191125",
    "pdfminer-six>=20260107",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.3",
]

[project.scripts]
invoice-intake-agent = "invoice_intake_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
